% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcfGuidedSpline.R
\name{atf.mgcv.spline}
\alias{atf.mgcv.spline}
\title{Autocorrelation Guided Spline Regression}
\usage{
atf.mgcv.spline(
  formula,
  data,
  df.range,
  spl.type = "cr",
  boxlag = 20,
  DEBUG = FALSE
)
}
\arguments{
\item{formula}{A regression formula with an unpenalized "bs" spline term to be optimized}

\item{data}{Data frame}

\item{df.range}{The range of df parameter to be optimized}

\item{spl.type}{The spline basis type: "cr" cubic, "tp" thin plate (default="cr")}

\item{boxlag}{The lag value for the Ljung Box test (default=20)}

\item{debug}{Debug mode: prints debug information when set to TRUE (default=FALSE).}
}
\value{
A list:
\itemize{
\item{"best.par"  : }{ Optimized df value}
\item{"best.pval" : }{ Ljung-Box pvalue corresponding to best.par}
\item{"best.fit"  : }{ GAM regression object corresponding to the best.par}
\item{"box.pvals" : }{ Vector of p-values relative to the range of parameters}
}
}
\description{
Spline optimization by residual autocorrelation analysis
}
\examples{
# Create a "iid + drift" signal
N <- 1000
drift <- sin(29.6*((1:N/N) + 0.2))/((1:N)/N + 0.2)
iid <- runif(n=N, min=-1.0, max=1.0)
y <- iid + drift
mydf <- data.frame(y=y,x=1:N)

# Add "random" covariates
mydf$cov1 <- sample(N)
mydf$cov2 <- sample(N)

# Run atf.mgcv.spline
fit <-   atf.mgcv.spline(y ~ bs(x) + cov1 + cov2, data=mydf, 3:100, DEBUG=TRUE)

# plot the "true" drift
lines(drift, col="blue")

# Plot LB p-values
plot(fit$box.pvals, type="l")
}
\keyword{atf.mgcv.spline}
