% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{dt.1}
\alias{dt.1}
\title{Example Dataset 1: Dumped oscillation + noise}
\format{
An object of class \code{data.frame} with 1000 rows and 2 columns.
}
\usage{
data(dt.1)
}
\description{
Example Dataset 1: Dumped oscillation + noise
}
\examples{
data(dt.1)
}
\keyword{datasets}
