% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{dt.2}
\alias{dt.2}
\title{Example Dataset 2: Stair + noise}
\format{
An object of class \code{data.frame} with 1000 rows and 2 columns.
}
\usage{
data(dt.2)
}
\description{
Example Dataset 2: Stair + noise
}
\examples{
data(dt.2)
}
\keyword{datasets}
