% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{dt.3}
\alias{dt.3}
\title{Example Dataset 3: Ramp + noise}
\format{
An object of class \code{data.frame} with 1000 rows and 2 columns.
}
\usage{
data(dt.3)
}
\description{
Example Dataset 3: Ramp + noise
}
\examples{
data(dt.3)
}
\keyword{datasets}
