% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{dt.4}
\alias{dt.4}
\title{Example Dataset 4: Harmonic + noise}
\format{
An object of class \code{data.frame} with 1000 rows and 2 columns.
}
\usage{
data(dt.4)
}
\description{
Example Dataset 4: Harmonic + noise
}
\examples{
data(dt.4)
}
\keyword{datasets}
