% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcfGuidedSpline.R
\name{atf.mgcv.spline}
\alias{atf.mgcv.spline}
\title{Autocorrelation Guided Spline Regression}
\usage{
atf.mgcv.spline(
  formula,
  dframe,
  df.range,
  spl.type = "cr",
  boxlag = 20,
  DEBUG = FALSE
)
}
\arguments{
\item{formula}{A regression formula with a spline term}

\item{dframe}{Data frame with output and exposure variable specified in formula}

\item{df.range}{The range of df parameter to be optimized}

\item{spl.type}{The spline basis type: "cr" cubic, "tp" thin plate (default="cr")}

\item{boxlag}{The lag value for the Ljung Box test (default=20)}

\item{debug}{Debug mode: prints debug information when set to TRUE (default=FALSE).}
}
\value{
A list (Ljung Box pvalues, best spline regression fit)
}
\description{
Spline optimization through residual autocorrelation analysis
}
\examples{
atf.mgcv.spline()
}
\keyword{atf.mgcv.spline}
