% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcfGuidedSpline.R
\name{atf.mgcv.spline}
\alias{atf.mgcv.spline}
\title{Autocorrelation Guided Spline Regression}
\usage{
atf.mgcv.spline(
  formula,
  dframe,
  df.range,
  spl.type = "cr",
  boxlag = 20,
  DEBUG = FALSE
)
}
\arguments{
\item{formula}{A regression formula with a spline term:
\itemize{
\item "bs": non-penalized
\item "s": penalized
}}

\item{dframe}{Data frame}

\item{df.range}{The range of df parameter to be optimized}

\item{spl.type}{The spline basis type: "cr" cubic, "tp" thin plate (default="cr")}

\item{boxlag}{The lag value for the Ljung Box test (default=20)}

\item{debug}{Debug mode: prints debug information when set to TRUE (default=FALSE).}
}
\value{
A list:
\itemize{
\item{"best.par"  : }{ Optimized parameter value}
\item{"best.pval" : }{ Ljung-Box pvalue corresponding to best.par}
\item{"box.pvals" : }{ Vector of p-values relative to the range of parameters}
\item{"best.fit"  : }{ GAM regression object corresponding to the best.par}
}
}
\description{
Spline optimization by residual autocorrelation analysis
}
\examples{
# Create a "iid + drift" signal
N <- 1000; T1<- N/8; T2 <- N/8
a1=2; a2=2
drift <- a1*sin(2.0*pi*(1:N)/T1) - a2*cos(2*pi*(1:N)/T2)
iid <- runif(n=N, min=-1.0, max=1.0)

# Drift + iid noise
y <- iid + drift
mydf <- data.frame(y=y,x=1:N)

# Add "covariates"
mydf$cov1 <- sample(N); mydf$cov2 <- sample(N)

# Run
myfit.1 <-   atf.mgcv.spline(y ~ bs(x) + cov2, dframe=mydf, 500:500, DEBUG=TRUE)
atf.mgcv.spline()
}
\keyword{atf.mgcv.spline}
